+++ frontend/src/backend/wails/index.ts
@@ -11,6 +11,7 @@ export { CreateFile as createFile } from "./go/main/App";
 export { openFile } from "./openFile";
 export { onOpenFile } from "./trigger";
 export { saveFileAs } from "./saveFileAs";
+export { ClipboardGetText as getClipboardText } from "./runtime";
 
 export { WindowSetTitle as setTitle } from "./runtime";
 export { SetIsDirty as setIsDirty } from "./go/main/App";