+++ frontend/src/drivers/supaplex/LevelLocalOptions.tsx
@@ -8,6 +8,8 @@ import { InlineTile } from "./InlineTile";
 import { TILE_HW_LAMP_R, TILE_HW_STRIPES, TILE_INFOTRON } from "./tiles-id";
 import { ISupaplexLevel } from "./types";
 import cl from "./LevelLocalOptions.module.scss";
+import { newSpecPortsDatabaseFromString } from "./specPortsDb";
+import { ISupaplexSpecPortDatabase } from "./internal";
 
 const compUsePlasma = {
   tile: constElement(<InlineTile tile={TILE_HW_STRIPES} />),
@@ -163,12 +165,24 @@ export const LevelLocalOptions = <L extends ISupaplexLevel>({
 // `use-infotrons-needed` is `in`
 //
 // `force-all-ports-special`
+
 const P_PLASMA = "2a";
+const P_USE_PLASMA = "use-plasma";
+const P_USE_PLASMA_LIMIT = "use-plasma-limit";
+const P_USE_PLASMA_TIME = "use-plasma-time";
+
 const P_ZONKER = "2b";
+const P_USE_ZONKER = "use-zonkers";
+
 const P_SERIAL_PORTS = "ps";
+const P_USE_SERIAL_PORTS = "use-serial-ports";
+
 const P_INFOTRONS_NEEDED = "in";
+const P_USE_INFOTRONS_NEEDED = "use-infotrons-needed";
+
 const P_FREEZE_ENEMIES = "fe";
 const P_PORTS_DB = "pd";
+
 export const applyLocalOptions = <L extends ISupaplexLevel>(
   level: L,
   url: URL,
@@ -198,4 +212,66 @@ export const applyLocalOptions = <L extends ISupaplexLevel>(
     p.set(P_PORTS_DB, level.specports.toString());
   }
   return url;
+};
+
+export const parseLocalOptions = <L extends ISupaplexLevel>(
+  url: URL,
+  level: L,
+): L => {
+  const p = url.searchParams;
+
+  let usePlasma = false;
+  let usePlasmaLimit: number | undefined;
+  let usePlasmaTime: number | undefined;
+  if (p.has(P_PLASMA)) {
+    level = level.setUsePlasma(true);
+    const m = p.get(P_PLASMA)!.match(/^(\d+)?x(\d+)?$/);
+    if (m) {
+      if (m[1] !== undefined) usePlasmaLimit = Number(m[1]);
+      if (m[2] !== undefined) usePlasmaTime = Number(m[2]);
+    }
+  } else {
+    usePlasma = p.has(P_USE_PLASMA);
+    if (p.has(P_USE_PLASMA_LIMIT)) {
+      usePlasmaLimit = parseInt(p.get(P_USE_PLASMA_LIMIT)!);
+    }
+    if (p.has(P_USE_PLASMA_TIME)) {
+      usePlasmaTime = parseInt(p.get(P_USE_PLASMA_TIME)!);
+    }
+  }
+
+  const useZonker = p.has(P_ZONKER) || p.has(P_USE_ZONKER);
+  const useSerialPorts = p.has(P_SERIAL_PORTS) || p.has(P_USE_SERIAL_PORTS);
+
+  let useInfotronsNeed: number | undefined;
+  {
+    const s = p.get(P_INFOTRONS_NEEDED) ?? p.get(P_USE_INFOTRONS_NEEDED);
+    if (s !== null) {
+      useInfotronsNeed = parseInt(s);
+    }
+  }
+
+  // TODO: byte
+  const useFreezeEnemies = p.has(P_FREEZE_ENEMIES);
+
+  let useSPDB: ISupaplexSpecPortDatabase | undefined;
+  {
+    const s = p.get(P_PORTS_DB);
+    if (s !== null) {
+      useSPDB = newSpecPortsDatabaseFromString(s);
+    }
+  }
+
+  return (
+    level
+      .setUsePlasma(usePlasma)
+      .setUsePlasmaLimit(usePlasmaLimit)
+      .setUsePlasmaTime(usePlasmaTime)
+      .setUseZonker(useZonker)
+      .setUseSerialPorts(useSerialPorts)
+      .setUseInfotronsNeeded(useInfotronsNeed)
+      // TODO: byte
+      .setInitialFreezeEnemies(useFreezeEnemies)
+      .updateSpecports((prev) => useSPDB ?? prev)
+  );
 };